refactor: sumcheck, test-vec-utils -> gkr (snark)
