+++ std/compress/io_test.go
@@ -1,9 +1,11 @@
-package compress
+package compress_test
 
 import (
 	"crypto/rand"
 	"errors"
 	"fmt"
+	"github.com/consensys/gnark/std/compress"
+	"github.com/consensys/gnark/std/compress/internal"
 	"math/big"
 	"testing"
 
@@ -14,7 +16,6 @@ import (
 	"github.com/consensys/gnark/frontend"
 	"github.com/consensys/gnark/frontend/cs/scs"
 	"github.com/consensys/gnark/profile"
-	test_vector_utils "github.com/consensys/gnark/std/internal/test_vectors_utils"
 	"github.com/consensys/gnark/test"
 	"github.com/stretchr/testify/assert"
 )
@@ -43,8 +44,8 @@ func TestShiftLeft(t *testing.T) {
 		}
 
 		assignment := shiftLeftCircuit{
-			Slice:       test_vector_utils.ToVariableSlice(b),
-			Shifted:     test_vector_utils.ToVariableSlice(shifted),
+			Slice:       internal.ToVariableSlice(b),
+			Shifted:     internal.ToVariableSlice(shifted),
 			ShiftAmount: shiftAmount,
 		}
 
@@ -77,7 +78,7 @@ func (c *shiftLeftCircuit) Define(api frontend.API) error {
 	if len(c.Slice) != len(c.Shifted) {
 		panic("witness length mismatch")
 	}
-	shifted := ShiftLeft(api, c.Slice, c.ShiftAmount)
+	shifted := compress.ShiftLeft(api, c.Slice, c.ShiftAmount)
 	if len(shifted) != len(c.Shifted) {
 		panic("wrong length")
 	}
@@ -94,14 +95,14 @@ func TestChecksumBytes(t *testing.T) {
 		_, err := rand.Read(b)
 		assert.NoError(t, err)
 
-		checksum := ChecksumPaddedBytes(b, len(b), hash.MIMC_BLS12_377.New(), fr.Bits)
+		checksum := compress.ChecksumPaddedBytes(b, len(b), hash.MIMC_BLS12_377.New(), fr.Bits)
 
 		circuit := checksumTestCircuit{
 			Bytes: make([]frontend.Variable, len(b)),
 		}
 
 		assignment := checksumTestCircuit{
-			Bytes: test_vector_utils.ToVariableSlice(b),
+			Bytes: internal.ToVariableSlice(b),
 			Sum:   checksum,
 		}
 
@@ -116,8 +117,8 @@ type checksumTestCircuit struct {
 }
 
 func (c *checksumTestCircuit) Define(api frontend.API) error {
-	Packed := append(Pack(api, c.Bytes, 8), len(c.Bytes))
-	return AssertChecksumEquals(api, Packed, c.Sum)
+	Packed := append(compress.Pack(api, c.Bytes, 8), len(c.Bytes))
+	return compress.AssertChecksumEquals(api, Packed, c.Sum)
 }
 
 func TestSetNumNbBits(t *testing.T) {
@@ -132,8 +133,8 @@ func TestSetNumNbBits(t *testing.T) {
 			test.WithCurves(ecc.BLS12_377), test.WithBackends(backend.PLONK),
 			test.WithValidAssignment(&testSetNumNbBitsCircuit{
 				increases: increases,
-				Words:     test_vector_utils.ToVariableSlice(words),
-				Nums:      test_vector_utils.ToVariableSlice(nums),
+				Words:     internal.ToVariableSlice(words),
+				Nums:      internal.ToVariableSlice(nums),
 			}))
 	}
 
@@ -195,7 +196,7 @@ func (c *testSetNumNbBitsCircuit) Define(api frontend.API) error {
 		return errors.New("must have as many steps as read values")
 	}
 	l := 1
-	nr := NewNumReader(api, c.Words, l, 1)
+	nr := compress.NewNumReader(api, c.Words, l, 1)
 	for i := range c.increases {
 		l += int(c.increases[i])
 		nr.SetNumNbBits(l)
